+++ adapters/Facebook/FacebookAdapter/GADFBNetworkExtras.m
@@ -16,4 +16,15 @@
 
 @implementation GADFBNetworkExtras
 
+- (instancetype)init {
+  self = [super init];
+  if (self) {
+    // Facebook's AdChoices icon is expandable and shows a background by default. If the
+    // publisher doesn't set one of these properties, we will use the default values.
+    self.adChoicesExpandable = YES;
+    self.adChoicesBackgroundShown = YES;
+  }
+  return self;
+}
+
 @end